backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
Pillow==10.4.0
opencv-python==4.10.0.84
numpy==1.26.4
pytesseract==0.3.13
scikit-learn==1.5.2
scipy==1.14.1
torch==2.4.1
torchvision==0.19.1
python-jose==3.3.0
aiofiles==24.1.0
pymupdf==1.24.10
